Refactor: useDebounceSearch hook 생성 및 사용
